requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==13.7
psycopg2-binary
pyowm
feedparser
geopy
speedtest-cli>=2.0.2
bs4
lxml
pytz
pyyaml
